+++ resources/views/layouts/partial/footer.blade.php
@@ -7,6 +7,7 @@
           <a href="{{route('site.terms')}}" class="text-dark p-2">{{__('site.terms')}}</a>
           <a href="{{route('site.privacy')}}" class="text-dark p-2">{{__('site.privacy')}}</a>
           <a href="{{route('site.language')}}" class="text-dark p-2">{{__('site.language')}}</a>
+          <a href="https://www.paypal.com/donate/?business=J7HKMWTQL7E8L&no_recurring=0&item_name=Contribua+para+o+crescimento+do+Pixelfed+Brasil%21&currency_code=BRL"  class="text-dark p-2" target="_blank" >Doar</a>
           @if(config_cache('instance.has_legal_notice'))
             <a href="/site/legal-notice" class="text-dark p-2">Legal Notice</a>
           @endif